Move `labels.py` out of `ome_zarr` subpackage`
