+++ fractal_tasks_core/tasks/napari_workflows_wrapper.py
@@ -27,11 +27,11 @@ from napari_workflows._io_yaml_v1 import load_workflow
 from pydantic.decorator import validate_arguments
 
 import fractal_tasks_core
+from fractal_tasks_core.labels import prepare_label_group
 from fractal_tasks_core.ngff import load_NgffImageMeta
 from fractal_tasks_core.ome_zarr.channels import get_channel_from_image_zarr
 from fractal_tasks_core.ome_zarr.input_models import NapariWorkflowsInput
 from fractal_tasks_core.ome_zarr.input_models import NapariWorkflowsOutput
-from fractal_tasks_core.ome_zarr.label_group import prepare_label_group
 from fractal_tasks_core.ome_zarr.pyramids import build_pyramid
 from fractal_tasks_core.ome_zarr.zattrs_utils import rescale_datasets
 from fractal_tasks_core.roi import check_valid_ROI_indices
