+++ tests/test_unit_write.py
@@ -6,7 +6,7 @@ import pytest
 import zarr
 from devtools import debug
 
-from fractal_tasks_core.ome_zarr.label_group import prepare_label_group
+from fractal_tasks_core.labels import prepare_label_group
 from fractal_tasks_core.tables.v1 import _write_elem_with_overwrite
 from fractal_tasks_core.write import open_zarr_group_with_overwrite
 from fractal_tasks_core.write import OverwriteNotAllowedError
